Adding One Ply analysis and pointed pairs fix
Analysis (if enabled) shows candidates you can not select as they will
lead to an invalid board state as well as candidates that will give
additional solved cells.

Previously, setting a value of a cell while running a rule would not
update the candidates of the box/row/col. This leads to incorrect
assumptions about the validity of the candidates. While not a perfect
solution, we record all pending updates to cell values until after the
completed pass of the rule. The start of the rule list will update the
candidates appropriately.

Long term, should either update candidates immediately after updating
values or separate the candidate inference logic from value logic.